Move all the merge operator tests to use DBPath
